organize files in push a bit more. fix compiling tests
