+++ Sources/MessagingPush/MessagingPush+AppDelegate.swift
@@ -1,0 +1,77 @@
+import Foundation
+#if canImport(UserNotifications) && canImport(UIKit)
+import UIKit
+import UserNotifications
+#endif
+
+#if canImport(UserNotifications) && canImport(UIKit)
+@available(iOSApplicationExtension, unavailable)
+public extension MessagingPush {
+    /**
+     A push notification was interacted with.
+
+     - returns: If the SDK called the completion handler for you indicating if the SDK took care of the request or not.
+     */
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void
+    ) -> Bool {
+        guard let implementation = implementation else {
+            completionHandler()
+            return false
+        }
+
+        return implementation.userNotificationCenter(center, didReceive: response,
+                                                     withCompletionHandler: completionHandler)
+    }
+}
+
+extension MessagingPushImplementation {
+    /**
+     A push notification was interacted with.
+
+     - returns: If the SDK called the completion handler for you indicating if the SDK took care of the request or not.
+     */
+    @available(iOSApplicationExtension, unavailable)
+    func userNotificationCenter(
+        _ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void
+    ) -> Bool {
+        if sdkConfigStore.config.autoTrackPushEvents {
+            var pushMetric = Metric.delivered
+
+            if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
+                pushMetric = Metric.opened
+            }
+
+            trackMetric(notificationContent: response.notification.request.content, event: pushMetric)
+        }
+
+        // Time to handle rich push notifications.
+        guard let pushContent = PushContent.parse(notificationContent: response.notification.request.content,
+                                                  jsonAdapter: jsonAdapter)
+        else {
+            // push does not contain a CIO rich payload, so end early
+            return false
+        }
+
+        cleanup(pushContent: pushContent)
+
+        switch response.actionIdentifier {
+        case UNNotificationDefaultActionIdentifier: // push notification was touched.
+            if let deepLinkurl = pushContent.deepLink {
+                UIApplication.shared.open(url: deepLinkurl)
+
+                completionHandler()
+
+                return true
+            }
+        default: break
+        }
+
+        return false
+    }
+}
+#endif
