+++ Sources/MessagingPush/MessagingPushImplementation.swift
@@ -154,52 +154,6 @@ internal class MessagingPushImplementation: MessagingPushInstance {
     }
 
     #if canImport(UserNotifications)
-    /**
-     A push notification was interacted with.
-
-     - returns: If the SDK called the completion handler for you indicating if the SDK took care of the request or not.
-     */
-    @available(iOSApplicationExtension, unavailable)
-    func userNotificationCenter(
-        _ center: UNUserNotificationCenter,
-        didReceive response: UNNotificationResponse,
-        withCompletionHandler completionHandler: @escaping () -> Void
-    ) -> Bool {
-        if sdkConfigStore.config.autoTrackPushEvents {
-            var pushMetric = Metric.delivered
-
-            if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
-                pushMetric = Metric.opened
-            }
-
-            trackMetric(notificationContent: response.notification.request.content, event: pushMetric)
-        }
-
-        // Time to handle rich push notifications.
-        guard let pushContent = PushContent.parse(notificationContent: response.notification.request.content,
-                                                  jsonAdapter: jsonAdapter)
-        else {
-            // push does not contain a CIO rich payload, so end early
-            return false
-        }
-
-        cleanup(pushContent: pushContent)
-
-        switch response.actionIdentifier {
-        case UNNotificationDefaultActionIdentifier: // push notification was touched.
-            if let deepLinkurl = pushContent.deepLink {
-                UIApplication.shared.open(url: deepLinkurl)
-
-                completionHandler()
-
-                return true
-            }
-        default: break
-        }
-
-        return false
-    }
-
     func trackMetric(
         notificationContent: UNNotificationContent,
         event: Metric
@@ -213,7 +167,7 @@ internal class MessagingPushImplementation: MessagingPushInstance {
         trackMetric(deliveryID: deliveryID, event: event, deviceToken: deviceToken)
     }
 
-    private func cleanup(pushContent: PushContent) {
+    internal func cleanup(pushContent: PushContent) {
         pushContent.cioAttachments.forEach { attachment in
             let localFilePath = attachment.url
