+++ Tests/MessagingPushAPN/APITest.swift
@@ -62,10 +62,9 @@ class MessagingPushAPNAPITest: UnitTest {
 
         _ = MessagingPush.shared.userNotificationCenter(.current(), didReceive: UNNotificationResponse.testInstance,
                                                         withCompletionHandler: {})
-        _ = mock.userNotificationCenter(UNUserNotificationCenter.current(),
-                                        didReceive: UNNotificationResponse.testInstance, withCompletionHandler: {})
         // cannot guarantee instance of protocol will have userNotificationCenter() function as that function is not available to app extensions.
-//        _ = instance.userNotificationCenter(.current(), didReceive: UNNotificationResponse.testInstance, withCompletionHandler: {})
+        // _ = mock.userNotificationCenter(UNUserNotificationCenter.current(), didReceive: UNNotificationResponse.testInstance, withCompletionHandler: {})
+        // _ = instance.userNotificationCenter(.current(), didReceive: UNNotificationResponse.testInstance, withCompletionHandler: {})
         #endif
     }
 }
